+++ HISTORY.md
@@ -22,6 +22,7 @@
 * objects: don't show some development-only blocks as search results
 * blocks: fixed a multi-line-text spec typo 
 * blocks: removed unused %month slot
+* blocks: removed unused %lst slot
 
 ### 2020-10-22
 * blocks: fixed UNDO/REDO for "extracted" (single) command blocks